// dist/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  admin: 'admin',
  pengajar: 'pengajar',
  santri: 'santri'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const StatusAbsensi = {
  hadir: 'hadir',
  izin: 'izin',
  alpha: 'alpha'
} as const

export type StatusAbsensi = (typeof StatusAbsensi)[keyof typeof StatusAbsensi]


export const StatusIzin = {
  menunggu: 'menunggu',
  disetujui: 'disetujui',
  ditolak: 'ditolak'
} as const

export type StatusIzin = (typeof StatusIzin)[keyof typeof StatusIzin]


export const StatusSubmission = {
  tepat_waktu: 'tepat_waktu',
  terlambat: 'terlambat'
} as const

export type StatusSubmission = (typeof StatusSubmission)[keyof typeof StatusSubmission]